requirements.txt
einops==0.8.0
transformers==4.47.1
vector_quantize_pytorch==1.19.2
pre-commit==3.6.0
deepspeed==0.14.4
sentencepiece==0.1.99
SwissArmyTransformer==0.4.12
omegaconf==2.3.0
pytorch_lightning==2.3.3
matplotlib==3.9.2
kornia==0.7.3
beartype==0.18.5
scipy==1.14.0
fiddle==0.3.0
imageio==2.34.2
imageio[ffmpeg]==0.6.0
wandb==0.17.5